fix : change array of selection
before : 'selectedCountryKey' list
after : 'selectedCountry' list
